requirements.txt
pyyaml>=6.0.1
click>=8.1

# Providers (optional)
openai>=1.40
google-generativeai>=0.7
boto3>=1.34
azure-identity>=1.16
azure-ai-inference>=1.0.0b8

# Training libs (CPU-only here; install torch with CUDA via conda/uv)
transformers>=4.44
accelerate>=0.33
mlflow>=2.15

# Storage
boto3>=1.34
google-cloud-storage>=2.18
azure-storage-blob>=12.22

# Dev
pytest>=8.3
ruff>=0.6
jsonschema>=4.23
